Move <Pagination /> component to common namespace
