+++ test/javascript/components/common/Pagination.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, fireEvent } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
-import { Pagination } from '../../../../app/javascript/components/mentoring/Pagination'
+import { Pagination } from '../../../../app/javascript/components/common/Pagination'
 
 test('clicking on "First" sets page to 1', () => {
   const setPage = jest.fn()
